whisper makes it to server and server modifies the msg to send
whisper sends from client1 to server, server detects and modifies the full message and attaches the sender's name to it. it then attempts to transmit the message to the target of the whisper.
change message detection to look for bools for /r and /n separately instead of in a stream from the same buffer
gave Node access to store_socket_t info through fd so it can begin message transmissions